Separacion de clases
Se separaron las clases en distintos archivos para evitar conflictos a la hora de hacer push
