+++ Clases/Cita.ts
@@ -1,0 +1,3 @@
+export class Cita{
+
+}
