+++ Clases/HistorialMedico.ts
@@ -1,0 +1,4 @@
+export class HistorialMedico{
+
+
+}
