+++ Clases/Paciente.ts
@@ -1,0 +1,3 @@
+export class Paciente {
+    
+}
