+++ Clases/Persona.ts
@@ -1,0 +1,9 @@
+
+export abstract class Persona{
+    constructor (protected id :string,protected Nombre:string,protected Apellido:string){
+
+    }
+    abstract modificar():void;
+    abstract consultar():void;
+
+}
