+++ Clases/RangoDeValores.ts
@@ -1,0 +1,3 @@
+export class RangoDeValores{
+
+}
